Upgrade code structure
