+++ src/command/CommandHandler.ts
@@ -1,0 +1,3 @@
+export default abstract class CommandHandler {
+    abstract handle(): void
+}
